pyproject.toml
[project]
name = "quick-stock"
version = "0.1.0"
description = "Simple inventory management system"
authors = [
    {name = "Skillica"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<=3.15"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "djlint (>=1.36.4,<2.0.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "django-browser-reload (>=1.21.0,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]
